petlink/requirements.txt
# --- FastAPI core ---
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.1
anyio==4.9.0
h11==0.16.0
sniffio==1.3.1

# --- Settings / config ---
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# --- Database ---
SQLAlchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.4
greenlet==3.2.3

psycopg2-binary==2.9.9

# --- Auth / security ---
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
PyJWT==2.10.1
cryptography==45.0.5
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1

# --- HTTP / utils ---
requests==2.32.3
urllib3==2.3.0
idna==3.10
charset-normalizer==3.4.1
certifi==2024.12.14
email_validator==2.2.0
dnspython==2.7.0

# --- Templates / misc (если реально используется) ---
Jinja2==3.1.5
MarkupSafe==3.0.2

# --- Dev tools ---
python-dotenv==1.1.1
flake8==7.1.1
pytest==8.3.3
pluggy==1.5.0
iniconfig==2.0.0
py==1.11.0
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
toml==0.10.2
packaging==24.2
six==1.17.0
